requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Authentication
PyJWT>=2.8.0
werkzeug>=2.3.0

# AI/ML Libraries
# Note: sentence-transformers and torch require significant resources
# Uncomment the following lines if your environment supports them:
# sentence-transformers>=2.2.0
# torch>=2.0.0

# Computer Vision
deepface>=0.0.90
opencv-python>=4.8.0
tensorflow>=2.13.0

# Audio Processing
speechrecognition>=3.10.0
pyttsx3>=2.90

# Communication
twilio>=8.0.0

# Data Visualization
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: For enhanced functionality
# scikit-learn>=1.3.0
# nltk>=3.8.0
# spacy>=3.6.0

